Handshake test runs successfully (and exits)
The problem seems to stem from the nested Systems that I'm creating when
the manager calls authorize. And, particularly, that the associated
auth-token is nil when it's time to shut those systems down.

I don't like this answer, but it's better than what I've been able to
come with previously. So I think I'll run with it until I either think
of something better or I have a better reason than "This smells bad" to
spend more time coming up with a better fix