Juggling pieces toward semblance of login
My dev environment is hosed. Last time I tried to actually run this, it
cratered my system. I think the actual problem has to do with
interactions between app armor and ubuntu kernel updates. This check-in
is really a back-up that belongs on a separate branch.

My apologies to future maintainers.