Initial check for incoming auth description broken
Partially because of semantic changes with the underlying 0mq API. I used to
just call recv! with :no-wait and deal with that not being there. Now it's
"properly" raising an EAGAIN.

But now, when I send the initial request to populate that info, I'm getting a
NPE. In the process of trying to debug that.

At this point, I've really split off into a totally different branch