Reconvince myself where negotiate-connection lives
It's part of world-manager because it doesn't make any sense for the
connection-manager to try to track the state of each connection it might
handle once its basic handshake is established (or rejected).

Get the bogus auth test to start and then sit around waiting for the
event loop to exit when it's failed because I really haven't written the
associated functionality.